Fixed vertex offset for complex meshes

+++ swMesh2XML/swMesh2XML/Triangle.cs
@@ -6,11 +6,11 @@ namespace swMesh2XML
 {
     class Triangle
     {
-        public byte v1;
-        public byte v2;
-        public byte v3;
+        public UInt16 v1;
+        public UInt16 v2;
+        public UInt16 v3;
 
-        public Triangle(byte v1, byte v2, byte v3)
+        public Triangle(UInt16 v1, UInt16 v2, UInt16 v3)
         {
             this.v1 = v1;
             this.v2 = v2;
